Formal specification for the Raft Pre-Vote mechanism
Introduce the Raft Pre-Vote mechanism as mentioned in section 4.2.3 of
the disseration, also further expanded on in section 9.6.

---

We introduce an explicit PreCandidate stage, the happy-path control flow
is now [Follower -> PreCandidate -> Candidate -> Leader]. PreCandidates
send out Pre-Votes to first check if any current quorum of nodes would be
willing to grant it votes should it run as a Candidate in the subsequent
term. If acknowledgement is received from a majority of nodes,
PreCandidates transition to the Candidate stage and call for an
election. If the PreCandidate receives responses from nodes at a higher
term (rejected Pre-Votes), it steps down to the Follower stage at the
specified term. If the PreCandidate does not receive a majority of
pre-votes, it will retry for the same next term after timing out (if it
does not receive heartbeats/append entries in the interim). Nodes can
give out pre-votes to multiple PreCandidates in the same term,
PreCandidates subsequently compete during the Candidate election stage.

Pre-Vote requests from nodes at a higher term cause the recipient term
to step down to a Follower stage at the specified term. This is similar
to the behaviour when Vote requests from higher termed Candidates are
received.